Start work over UserDashboard trade functionality

+++ TopTrade/Services/TopTrade.Services.Data/User/AlphaVantageApiClientService.cs
@@ -68,8 +68,8 @@
             {
                 Ticker = stockResult.Symbol,
                 Price = stockResult.Price,
-                Change = stockResult.Change,
-                ChangePercent = stockResult.ChangePercent,
+                Change = (double)stockResult.Change,
+                ChangePercent = (double)stockResult.ChangePercent,
             };
 
             return stock;
