+++ TopTrade/Web/TopTrade.Web.ViewModels/User/StockViewModel.cs
@@ -12,8 +12,8 @@
 
         public decimal Price { get; set; }
 
-        public decimal Change { get; set; }
+        public double Change { get; set; }
 
-        public decimal ChangePercent { get; set; }
+        public double ChangePercent { get; set; }
     }
 }
